+++ src/core/server/extensions/playerFunctions/inventory.ts
@@ -678,7 +678,7 @@ function handleSwapOrStack(player: alt.Player, selectedSlot: string, endSlot: st
     const endArray: Array<Item> = selectedSlotName === endSlotName ? selectedArray : player.data[endSlotName];
 
     // Do Not Stack. Swap Instead.
-    if (selectItem.item.name !== endItem.item.name) {
+    if (selectItem.item.name !== endItem.item.name || selectItem.item.rarity !== selectItem.item.rarity) {
         // Need to verify that each slot follows the rules for the slot it is going into.
         // Handles rules for the end item slot.
         if (!allItemRulesValid(selectItem.item, { name: endSlotName }, newEndSlot)) {
@@ -958,6 +958,10 @@ function stackInventoryItem(player: alt.Player, item: Item): boolean {
     const existingItem = isInInventory(player, item);
 
     if (!existingItem) {
+        return false;
+    }
+
+    if (item.rarity !== player.data.inventory[existingItem.index].rarity) {
         return false;
     }